+++ src/main.js
@@ -108,6 +108,28 @@ Vue.mixin({
       if (resource.charAt(0) == '/') resource = resource.slice(1)
       return this.$localhost + resource
     },
+    $filterLinksAndImages(el){
+      // Check if this is a comment
+      if (el.nodeType !== 8){
+        let links = el.getElementsByTagName('A')
+        for(let i = 0;i < links.length; i++) {
+          links[i].href = this.$filterURL(links[i].href)
+        }
+
+        //Normally, this process would be handled by the MediaEmbed component. Gotta get the behaviour into all them images somehow!
+        let images = [...el.getElementsByTagName('IMG'), ...el.getElementsByTagName('VIDEO')]
+        for(let i = 0;i < images.length; i++) {
+          images[i].src = this.$mspaURL(images[i].src)
+          if (images[i].tagName == 'IMG') {
+            images[i].ondragstart = (e) => {
+              e.preventDefault()
+              e.dataTransfer.effectAllowed = 'copy'
+              require('electron').ipcRenderer.send('ondragstart', this.$mspaFileStream(images[i].src))
+            }
+          }
+        }
+      }
+    },
     $getStory(pageNumber){
       pageNumber = parseInt(pageNumber) || pageNumber
